Add book_followup cli cmd
to be able to book a follow up appointment 21 days after the given
booking with that very same data beside the date.

It is checked that the user given in booked_by attribute did not book
any appointment yet.

Free slots are searched 21 days after the start_date_time of the
given booking, if none are found on that same day, no booking happens.

The appointment is first claimed and then booked.